♻️ Refactor: Move the location of the sendMessageLocationUpdate function to the WebSocketService class
